requirement_pt.txt
numpy
pandas
sqlalchemy
psycopg2-binary
spacy
pyyaml
requests
tqdm
pillow
wordcloud
geopy
geopandas
scikit-learn
torchtext
torchvision
transformers
tensorboard
gensim
emoji
torchinfo
